Re-establishing logging after completed job
