+++ pilot/control/job.py
@@ -34,7 +34,7 @@ from pilot.util.common import should_abort
 from pilot.util.constants import PILOT_PRE_GETJOB, PILOT_POST_GETJOB, PILOT_KILL_SIGNAL, LOG_TRANSFER_NOT_DONE, \
     LOG_TRANSFER_IN_PROGRESS, LOG_TRANSFER_DONE, LOG_TRANSFER_FAILED, SERVER_UPDATE_TROUBLE, SERVER_UPDATE_FINAL, \
     SERVER_UPDATE_UPDATING, SERVER_UPDATE_NOT_DONE
-from pilot.util.filehandling import get_files, tail, is_json, copy, remove, read_file, write_json
+from pilot.util.filehandling import get_files, tail, is_json, copy, remove, read_file, write_json, establish_logging
 from pilot.util.harvester import request_new_jobs, remove_job_request_file, parse_job_definition_file, \
     is_harvester_mode, get_worker_attributes_file, publish_work_report, get_event_status_file, \
     publish_stageout_files
@@ -1345,6 +1345,12 @@ def retrieve(queues, traces, args):
                 while not args.graceful_stop.is_set():
                     if has_job_completed(queues):
                         logger.info('ready for new job')
+
+                        # re-establish logging
+                        logging.info('pilot has finished for previous job - re-establishing logging')
+                        logging.shutdown()
+                        establish_logging(args)
+
                         getjob_requests = 0
                         break
                     time.sleep(0.5)
